requirements.txt
rs3clans
discord.py
python-dotenv